refactor: move the OGSketch from agg_iterator.gen.go to agg_iterator.go
