+++ engine/executor/agg_iterator.gen.go
@@ -21,11 +21,9 @@ limitations under the License.
 package executor
 
 import (
-	"bytes"
 	"sort"
 
 	"github.com/openGemini/openGemini/engine/hybridqp"
-	"github.com/openGemini/openGemini/lib/util/lifted/influx/query"
 )
 
 type FloatTimeColFloatIterator struct {
@@ -3020,768 +3018,6 @@ func linearFloat(windowTime, previousTime, nextTime int64, previousValue, nextVa
 	return m*x + b
 }
 
-type FloatIntegralItem struct {
-	sameTag      bool
-	sameInterval bool
-	pointNum     int
-	sum          float64
-	window       struct {
-		start int64
-		end   int64
-	}
-	prev     *Point[float64]
-	time     []int64
-	value    []float64
-	interval hybridqp.Interval
-	opt      *query.ProcessorOptions
-}
-
-func NewFloatIntegralItem(interval hybridqp.Interval, opt *query.ProcessorOptions) *FloatIntegralItem {
-	return &FloatIntegralItem{interval: interval, prev: newPoint[float64](), opt: opt, sameTag: false}
-}
-
-func (f *FloatIntegralItem) CalculateUnit(index int, time int64, value float64) {
-	if f.prev.time == time {
-		f.prev.Set(index, time, value)
-	} else {
-		f.sum += 0.5 * (value + f.prev.value) * float64(time-f.prev.time) / float64(f.interval.Duration)
-		f.prev.Set(index, time, value)
-	}
-}
-
-func (f *FloatIntegralItem) StartNewInterval(time int64) {
-	f.value = append(f.value, f.sum)
-	if f.opt.Interval.IsZero() {
-		f.time = append(f.time, 0)
-	} else {
-		f.time = append(f.time, f.window.start)
-		if f.opt.Ascending {
-			f.window.start, f.window.end = f.opt.Window(time)
-		} else {
-			f.window.end, f.window.start = f.opt.Window(time)
-		}
-	}
-	f.sum = 0.0
-}
-
-func (f *FloatIntegralItem) doNullWindow(sameInterval, sameTag bool) {
-	if !f.sameTag {
-		if f.pointNum > 1 {
-			f.value = append(f.value, f.sum)
-			if f.opt.Interval.IsZero() {
-				f.time = append(f.time, 0)
-			} else {
-				f.time = append(f.time, f.window.start)
-			}
-		}
-		f.sum = 0.0
-		f.pointNum = 0
-	} else {
-		f.sameInterval = sameInterval
-		f.sameTag = sameTag
-	}
-}
-
-func (f *FloatIntegralItem) AppendItemFastFunc(c Chunk, ordinal int, start, end int, sameInterval bool, sameTag bool) {
-	// fast path
-	col := c.Column(ordinal)
-	time := c.Time()[start:end]
-	value := col.FloatValues()[start:end]
-
-	// process the first point
-	if f.prev.isNil {
-		f.prev.Set(0, time[0], float64(value[0]))
-		if !f.opt.Interval.IsZero() {
-			if f.opt.Ascending {
-				f.window.start, f.window.end = f.opt.Window(time[0])
-			} else {
-				f.window.end, f.window.start = f.opt.Window(time[0])
-			}
-		}
-	} else {
-		// process the last point of front window and the first point of this window
-		if !f.sameTag { // not sametag
-			if !(f.pointNum == 1 && f.prev.time == f.window.start) && !(f.pointNum == 0) {
-				f.StartNewInterval(time[0])
-			} else {
-				if !f.opt.Interval.IsZero() {
-					if f.opt.Ascending {
-						f.window.start, f.window.end = f.opt.Window(time[0])
-					} else {
-						f.window.end, f.window.start = f.opt.Window(time[0])
-					}
-				}
-				f.sum = 0.0
-			}
-			f.prev.Set(0, time[0], float64(value[0]))
-		} else if !f.sameInterval { // sametag not sameinterval
-			if f.prev.time != f.window.end && !f.opt.Interval.IsZero() {
-				value := linearFloat(f.window.end, f.prev.time, time[0], f.prev.value, float64(value[0]))
-				f.sum += 0.5 * (value + f.prev.value) * float64(f.window.end-f.prev.time) / float64(f.interval.Duration)
-
-				f.prev.value = value
-				f.prev.time = f.window.end
-			}
-			f.StartNewInterval(time[0])
-			f.CalculateUnit(0, time[0], float64(value[0]))
-		} else { // sametag sameinterval
-			f.CalculateUnit(0, time[0], float64(value[0]))
-		}
-	}
-	// process the rest ponints
-	for i := 1; i < len(time); i++ {
-		f.CalculateUnit(i, time[i], float64(value[i]))
-	}
-	f.pointNum = end - start
-	f.sameTag = sameTag
-	f.sameInterval = sameInterval
-}
-
-func (f *FloatIntegralItem) AppendItemSlowFunc(c Chunk, ordinal int, vs, ve int, sameInterval, sameTag bool) {
-	// slow path
-	col := c.Column(ordinal)
-	getTimeIndex := col.GetTimeIndex
-	if vs == ve {
-		f.doNullWindow(sameInterval, sameTag)
-		return
-	}
-
-	// process the first point
-	if f.prev.isNil {
-		f.prev.Set(0, c.TimeByIndex(getTimeIndex(vs)), float64(col.FloatValues()[vs]))
-		if !f.opt.Interval.IsZero() {
-			if f.opt.Ascending {
-				f.window.start, f.window.end = f.opt.Window(c.TimeByIndex(getTimeIndex(vs)))
-			} else {
-				f.window.end, f.window.start = f.opt.Window(c.TimeByIndex(getTimeIndex(vs)))
-			}
-		}
-	} else {
-		// process the last point of front window and the first point of this window
-		if !f.sameTag {
-			if !(f.pointNum == 1 && f.prev.time == f.window.start) && !(f.pointNum == 0) {
-				f.StartNewInterval(c.TimeByIndex(getTimeIndex(vs)))
-			} else {
-				if !f.opt.Interval.IsZero() {
-					if f.opt.Ascending {
-						f.window.start, f.window.end = f.opt.Window(c.TimeByIndex(getTimeIndex(vs)))
-					} else {
-						f.window.end, f.window.start = f.opt.Window(c.TimeByIndex(getTimeIndex(vs)))
-					}
-				}
-				f.sum = 0.0
-			}
-			f.prev.Set(0, c.TimeByIndex(getTimeIndex(vs)), float64(col.FloatValues()[vs]))
-		} else if !f.sameInterval {
-			if f.prev.time != f.window.end && !f.opt.Interval.IsZero() {
-				value := linearFloat(f.window.end, f.prev.time, c.TimeByIndex(getTimeIndex(vs)), f.prev.value, float64(col.FloatValues()[vs]))
-				f.sum += 0.5 * (value + f.prev.value) * float64(f.window.end-f.prev.time) / float64(f.interval.Duration)
-
-				f.prev.value = value
-				f.prev.time = f.window.end
-			}
-			f.StartNewInterval(c.TimeByIndex(getTimeIndex(vs)))
-			f.CalculateUnit(0, c.TimeByIndex(getTimeIndex(vs)), float64(col.FloatValues()[vs]))
-		} else {
-			f.CalculateUnit(0, c.TimeByIndex(getTimeIndex(vs)), float64(col.FloatValues()[vs]))
-		}
-	}
-	// process the rest ponints
-	for i := vs + 1; i < ve; i++ {
-		f.CalculateUnit(i-vs, c.TimeByIndex(getTimeIndex(i)), float64(col.FloatValues()[i]))
-	}
-	f.pointNum = ve - vs
-	f.sameTag = sameTag
-	f.sameInterval = sameInterval
-}
-
-func (f *FloatIntegralItem) AppendItem(c Chunk, ordinal int, start, end int, sameInterval, sameTag bool) {
-	if c.Column(ordinal).NilCount() == 0 {
-		f.AppendItemFastFunc(c, ordinal, start, end, sameInterval, sameTag)
-		return
-	}
-	f.AppendItemSlowFunc(c, ordinal, start, end, sameInterval, sameTag)
-}
-
-func (f *FloatIntegralItem) Reset() {
-	f.time = f.time[:0]
-	f.value = f.value[:0]
-}
-
-func (f *FloatIntegralItem) Len() int {
-	return len(f.time)
-}
-
-func (f *FloatIntegralItem) Nil() bool {
-	return f.prev.isNil
-}
-
-type IntegerIntegralItem struct {
-	sameTag      bool
-	sameInterval bool
-	pointNum     int
-	sum          float64
-	window       struct {
-		start int64
-		end   int64
-	}
-	prev     *Point[float64]
-	time     []int64
-	value    []float64
-	interval hybridqp.Interval
-	opt      *query.ProcessorOptions
-}
-
-func NewIntegerIntegralItem(interval hybridqp.Interval, opt *query.ProcessorOptions) *IntegerIntegralItem {
-	return &IntegerIntegralItem{interval: interval, prev: newPoint[float64](), opt: opt, sameTag: false}
-}
-
-func (f *IntegerIntegralItem) CalculateUnit(index int, time int64, value float64) {
-	if f.prev.time == time {
-		f.prev.Set(index, time, value)
-	} else {
-		f.sum += 0.5 * (value + f.prev.value) * float64(time-f.prev.time) / float64(f.interval.Duration)
-		f.prev.Set(index, time, value)
-	}
-}
-
-func (f *IntegerIntegralItem) StartNewInterval(time int64) {
-	f.value = append(f.value, f.sum)
-	if f.opt.Interval.IsZero() {
-		f.time = append(f.time, 0)
-	} else {
-		f.time = append(f.time, f.window.start)
-		if f.opt.Ascending {
-			f.window.start, f.window.end = f.opt.Window(time)
-		} else {
-			f.window.end, f.window.start = f.opt.Window(time)
-		}
-	}
-	f.sum = 0.0
-}
-
-func (f *IntegerIntegralItem) doNullWindow(sameInterval, sameTag bool) {
-	if !f.sameTag {
-		if f.pointNum > 1 {
-			f.value = append(f.value, f.sum)
-			if f.opt.Interval.IsZero() {
-				f.time = append(f.time, 0)
-			} else {
-				f.time = append(f.time, f.window.start)
-			}
-		}
-		f.sum = 0.0
-		f.pointNum = 0
-	} else {
-		f.sameInterval = sameInterval
-		f.sameTag = sameTag
-	}
-}
-
-func (f *IntegerIntegralItem) AppendItemFastFunc(c Chunk, ordinal int, start, end int, sameInterval bool, sameTag bool) {
-	// fast path
-	col := c.Column(ordinal)
-	time := c.Time()[start:end]
-	value := col.IntegerValues()[start:end]
-
-	// process the first point
-	if f.prev.isNil {
-		f.prev.Set(0, time[0], float64(value[0]))
-		if !f.opt.Interval.IsZero() {
-			if f.opt.Ascending {
-				f.window.start, f.window.end = f.opt.Window(time[0])
-			} else {
-				f.window.end, f.window.start = f.opt.Window(time[0])
-			}
-		}
-	} else {
-		// process the last point of front window and the first point of this window
-		if !f.sameTag { // not sametag
-			if !(f.pointNum == 1 && f.prev.time == f.window.start) && !(f.pointNum == 0) {
-				f.StartNewInterval(time[0])
-			} else {
-				if !f.opt.Interval.IsZero() {
-					if f.opt.Ascending {
-						f.window.start, f.window.end = f.opt.Window(time[0])
-					} else {
-						f.window.end, f.window.start = f.opt.Window(time[0])
-					}
-				}
-				f.sum = 0.0
-			}
-			f.prev.Set(0, time[0], float64(value[0]))
-		} else if !f.sameInterval { // sametag not sameinterval
-			if f.prev.time != f.window.end && !f.opt.Interval.IsZero() {
-				value := linearFloat(f.window.end, f.prev.time, time[0], f.prev.value, float64(value[0]))
-				f.sum += 0.5 * (value + f.prev.value) * float64(f.window.end-f.prev.time) / float64(f.interval.Duration)
-
-				f.prev.value = value
-				f.prev.time = f.window.end
-			}
-			f.StartNewInterval(time[0])
-			f.CalculateUnit(0, time[0], float64(value[0]))
-		} else { // sametag sameinterval
-			f.CalculateUnit(0, time[0], float64(value[0]))
-		}
-	}
-	// process the rest ponints
-	for i := 1; i < len(time); i++ {
-		f.CalculateUnit(i, time[i], float64(value[i]))
-	}
-	f.pointNum = end - start
-	f.sameTag = sameTag
-	f.sameInterval = sameInterval
-}
-
-func (f *IntegerIntegralItem) AppendItemSlowFunc(c Chunk, ordinal int, vs, ve int, sameInterval, sameTag bool) {
-	// slow path
-	col := c.Column(ordinal)
-	getTimeIndex := col.GetTimeIndex
-	if vs == ve {
-		f.doNullWindow(sameInterval, sameTag)
-		return
-	}
-
-	// process the first point
-	if f.prev.isNil {
-		f.prev.Set(0, c.TimeByIndex(getTimeIndex(vs)), float64(col.IntegerValues()[vs]))
-		if !f.opt.Interval.IsZero() {
-			if f.opt.Ascending {
-				f.window.start, f.window.end = f.opt.Window(c.TimeByIndex(getTimeIndex(vs)))
-			} else {
-				f.window.end, f.window.start = f.opt.Window(c.TimeByIndex(getTimeIndex(vs)))
-			}
-		}
-	} else {
-		// process the last point of front window and the first point of this window
-		if !f.sameTag {
-			if !(f.pointNum == 1 && f.prev.time == f.window.start) && !(f.pointNum == 0) {
-				f.StartNewInterval(c.TimeByIndex(getTimeIndex(vs)))
-			} else {
-				if !f.opt.Interval.IsZero() {
-					if f.opt.Ascending {
-						f.window.start, f.window.end = f.opt.Window(c.TimeByIndex(getTimeIndex(vs)))
-					} else {
-						f.window.end, f.window.start = f.opt.Window(c.TimeByIndex(getTimeIndex(vs)))
-					}
-				}
-				f.sum = 0.0
-			}
-			f.prev.Set(0, c.TimeByIndex(getTimeIndex(vs)), float64(col.IntegerValues()[vs]))
-		} else if !f.sameInterval {
-			if f.prev.time != f.window.end && !f.opt.Interval.IsZero() {
-				value := linearFloat(f.window.end, f.prev.time, c.TimeByIndex(getTimeIndex(vs)), f.prev.value, float64(col.IntegerValues()[vs]))
-				f.sum += 0.5 * (value + f.prev.value) * float64(f.window.end-f.prev.time) / float64(f.interval.Duration)
-
-				f.prev.value = value
-				f.prev.time = f.window.end
-			}
-			f.StartNewInterval(c.TimeByIndex(getTimeIndex(vs)))
-			f.CalculateUnit(0, c.TimeByIndex(getTimeIndex(vs)), float64(col.IntegerValues()[vs]))
-		} else {
-			f.CalculateUnit(0, c.TimeByIndex(getTimeIndex(vs)), float64(col.IntegerValues()[vs]))
-		}
-	}
-	// process the rest ponints
-	for i := vs + 1; i < ve; i++ {
-		f.CalculateUnit(i-vs, c.TimeByIndex(getTimeIndex(i)), float64(col.IntegerValues()[i]))
-	}
-	f.pointNum = ve - vs
-	f.sameTag = sameTag
-	f.sameInterval = sameInterval
-}
-
-func (f *IntegerIntegralItem) AppendItem(c Chunk, ordinal int, start, end int, sameInterval, sameTag bool) {
-	if c.Column(ordinal).NilCount() == 0 {
-		f.AppendItemFastFunc(c, ordinal, start, end, sameInterval, sameTag)
-		return
-	}
-	f.AppendItemSlowFunc(c, ordinal, start, end, sameInterval, sameTag)
-}
-
-func (f *IntegerIntegralItem) Reset() {
-	f.time = f.time[:0]
-	f.value = f.value[:0]
-}
-
-func (f *IntegerIntegralItem) Len() int {
-	return len(f.time)
-}
-
-func (f *IntegerIntegralItem) Nil() bool {
-	return f.prev.isNil
-}
-
-type FloatColFloatIntegralIterator struct {
-	isSingleCall bool
-	inOrdinal    int
-	outOrdinal   int
-	nilCount     int
-	buf          *FloatIntegralItem
-}
-
-func NewFloatColFloatIntegralIterator(
-	isSingleCall bool, inOrdinal, outOrdinal int, rowDataType hybridqp.RowDataType, interval hybridqp.Interval,
-	opt *query.ProcessorOptions,
-) *FloatColFloatIntegralIterator {
-	r := &FloatColFloatIntegralIterator{
-		buf:          NewFloatIntegralItem(interval, opt),
-		isSingleCall: isSingleCall,
-		inOrdinal:    inOrdinal,
-		outOrdinal:   outOrdinal,
-		nilCount:     0,
-	}
-	return r
-}
-
-func (r *FloatColFloatIntegralIterator) processFirstWindow(
-	inChunk Chunk, haveMultiInterval, sameInterval, sameTag bool, start, end int,
-) {
-	if haveMultiInterval {
-		r.buf.AppendItem(inChunk, r.inOrdinal, start, end, false, sameTag)
-	} else {
-		r.buf.AppendItem(inChunk, r.inOrdinal, start, end, sameInterval, sameTag)
-	}
-	if !sameTag && (end-start) == 1 && r.buf.prev.time == r.buf.window.start {
-		r.nilCount++
-	}
-}
-
-func (r *FloatColFloatIntegralIterator) processLastWindow(
-	inChunk, outChunk Chunk, start, end, tagIdx int,
-) {
-	r.buf.AppendItem(inChunk, r.inOrdinal, start, end, true, true)
-	if r.buf.Len() > 0 {
-		r.appendPrevItem(inChunk, outChunk, tagIdx)
-		r.buf.Reset()
-		if !r.isSingleCall {
-			outChunk.Column(r.outOrdinal).AppendManyNil(r.nilCount)
-			r.nilCount = 0
-		}
-	}
-}
-
-func (r *FloatColFloatIntegralIterator) processMiddleWindow(
-	inChunk, outChunk Chunk, start, end, tagIdx int, sameTag bool,
-) {
-	r.buf.AppendItem(inChunk, r.inOrdinal, start, end, false, sameTag)
-	if !sameTag && (end-start) == 1 && r.buf.prev.time == r.buf.window.start {
-		r.nilCount++
-	}
-	if r.buf.Len() > 0 {
-		r.appendPrevItem(inChunk, outChunk, tagIdx)
-		r.buf.Reset()
-		if !r.isSingleCall {
-			outChunk.Column(r.outOrdinal).AppendManyNil(r.nilCount)
-			r.nilCount = 0
-		}
-	}
-}
-
-func (r *FloatColFloatIntegralIterator) appendPrevItem(
-	inChunk, outChunk Chunk, tagIdx int,
-) {
-	outColumn := outChunk.Column(r.outOrdinal)
-	if r.isSingleCall {
-		for j := range r.buf.time {
-			outChunk.AppendTime(r.buf.time[j])
-			outColumn.AppendFloatValue(r.buf.value[j])
-			outColumn.AppendNotNil()
-		}
-		outChunk.AppendIntervalIndex(outChunk.Len() - r.buf.Len())
-		if !(outChunk.TagLen() > 0 && bytes.Equal(inChunk.Tags()[tagIdx].Subset(r.buf.opt.Dimensions),
-			outChunk.Tags()[outChunk.TagLen()-1].Subset(r.buf.opt.Dimensions))) {
-			outChunk.AppendTagsAndIndex(inChunk.Tags()[tagIdx], outChunk.IntervalIndex()[outChunk.IntervalLen()-1])
-		}
-		return
-	}
-
-	for j := range r.buf.time {
-		outColumn.AppendFloatValue(r.buf.value[j])
-		outColumn.AppendNotNil()
-	}
-}
-
-func (r *FloatColFloatIntegralIterator) appendLastItem(
-	inChunk, outChunk Chunk, tagIdx int,
-) {
-	outColumn := outChunk.Column(r.outOrdinal)
-	if r.isSingleCall {
-		if r.buf.opt.Interval.IsZero() {
-			outChunk.AppendTime(0)
-		} else {
-			outChunk.AppendTime(r.buf.window.start)
-		}
-		outColumn.AppendFloatValue(r.buf.sum)
-		outColumn.AppendNotNil()
-		outChunk.AppendIntervalIndex(outChunk.Len() - 1)
-		if !(outChunk.TagLen() > 0 && bytes.Equal(inChunk.Tags()[tagIdx].Subset(r.buf.opt.Dimensions),
-			outChunk.Tags()[outChunk.TagLen()-1].Subset(r.buf.opt.Dimensions))) {
-			outChunk.AppendTagsAndIndex(inChunk.Tags()[tagIdx], outChunk.IntervalIndex()[outChunk.IntervalLen()-1])
-		}
-		return
-	}
-
-	outColumn.AppendFloatValue(r.buf.sum)
-	outColumn.AppendNotNil()
-}
-
-func (r *FloatColFloatIntegralIterator) Next(ie *IteratorEndpoint, p *IteratorParams) {
-	inChunk, outChunk := ie.InputPoint.Chunk, ie.OutputPoint.Chunk
-	inColumn := inChunk.Column(r.inOrdinal)
-	if inColumn.IsEmpty() && r.buf.Len() > 0 {
-		return
-	}
-
-	var start, end int
-	firstIndex, lastIndex := 0, len(inChunk.IntervalIndex())-1
-	lastTagIndex := len(inChunk.TagIndex()) - 1
-	intervali := 0
-	var subIntervalIndexEnd int
-	outColumn := outChunk.Column(r.outOrdinal)
-	for i, _ := range inChunk.TagIndex() {
-		if i == lastTagIndex {
-			subIntervalIndexEnd = inChunk.NumberOfRows()
-		} else {
-			subIntervalIndexEnd = inChunk.TagIndex()[i+1]
-		}
-		for inChunk.IntervalIndex()[intervali] < subIntervalIndexEnd {
-			start = inChunk.IntervalIndex()[intervali]
-			if intervali < lastIndex {
-				end = inChunk.IntervalIndex()[intervali+1]
-			} else {
-				end = subIntervalIndexEnd
-			}
-			var sametag bool
-			if end == inChunk.NumberOfRows() {
-				sametag = !(end == subIntervalIndexEnd) || p.sameTag
-			} else {
-				sametag = !(end == subIntervalIndexEnd)
-			}
-			if !r.isSingleCall {
-				start, end = inColumn.GetRangeValueIndexV2(start, end)
-				if start == end {
-					if r.buf.Nil() {
-						outColumn.AppendNil()
-						intervali++
-						if intervali >= len(inChunk.IntervalIndex()) {
-							break
-						}
-						continue
-					} else {
-						r.nilCount++
-					}
-				}
-			}
-			if intervali == firstIndex && r.buf.sameInterval {
-				r.processFirstWindow(inChunk, firstIndex != lastIndex, p.sameInterval, sametag, start, end)
-			} else if intervali == lastIndex && p.sameInterval {
-				r.processLastWindow(inChunk, outChunk, start, end, i)
-			} else {
-				r.processMiddleWindow(inChunk, outChunk, start, end, i, sametag)
-			}
-			intervali++
-			if intervali >= len(inChunk.IntervalIndex()) {
-				break
-			}
-		}
-	}
-	if p.lastChunk {
-		if !(r.buf.pointNum == 1 && r.buf.prev.time == r.buf.window.start) && !(r.buf.pointNum == 0) {
-			r.appendLastItem(inChunk, outChunk, inChunk.TagLen()-1)
-		}
-		if !r.isSingleCall {
-			outColumn.AppendManyNil(r.nilCount)
-			r.nilCount = 0
-		}
-	}
-}
-
-type IntegerColFloatIntegralIterator struct {
-	isSingleCall bool
-	inOrdinal    int
-	outOrdinal   int
-	nilCount     int
-	buf          *IntegerIntegralItem
-}
-
-func NewIntegerColFloatIntegralIterator(
-	isSingleCall bool, inOrdinal, outOrdinal int, rowDataType hybridqp.RowDataType, interval hybridqp.Interval,
-	opt *query.ProcessorOptions,
-) *IntegerColFloatIntegralIterator {
-	r := &IntegerColFloatIntegralIterator{
-		buf:          NewIntegerIntegralItem(interval, opt),
-		isSingleCall: isSingleCall,
-		inOrdinal:    inOrdinal,
-		outOrdinal:   outOrdinal,
-		nilCount:     0,
-	}
-	return r
-}
-
-func (r *IntegerColFloatIntegralIterator) processFirstWindow(
-	inChunk Chunk, haveMultiInterval, sameInterval, sameTag bool, start, end int,
-) {
-	if haveMultiInterval {
-		r.buf.AppendItem(inChunk, r.inOrdinal, start, end, false, sameTag)
-	} else {
-		r.buf.AppendItem(inChunk, r.inOrdinal, start, end, sameInterval, sameTag)
-	}
-	if !sameTag && (end-start) == 1 && r.buf.prev.time == r.buf.window.start {
-		r.nilCount++
-	}
-}
-
-func (r *IntegerColFloatIntegralIterator) processLastWindow(
-	inChunk, outChunk Chunk, start, end, tagIdx int,
-) {
-	r.buf.AppendItem(inChunk, r.inOrdinal, start, end, true, true)
-	if r.buf.Len() > 0 {
-		r.appendPrevItem(inChunk, outChunk, tagIdx)
-		r.buf.Reset()
-		if !r.isSingleCall {
-			outChunk.Column(r.outOrdinal).AppendManyNil(r.nilCount)
-			r.nilCount = 0
-		}
-	}
-}
-
-func (r *IntegerColFloatIntegralIterator) processMiddleWindow(
-	inChunk, outChunk Chunk, start, end, tagIdx int, sameTag bool,
-) {
-	r.buf.AppendItem(inChunk, r.inOrdinal, start, end, false, sameTag)
-	if !sameTag && (end-start) == 1 && r.buf.prev.time == r.buf.window.start {
-		r.nilCount++
-	}
-	if r.buf.Len() > 0 {
-		r.appendPrevItem(inChunk, outChunk, tagIdx)
-		r.buf.Reset()
-		if !r.isSingleCall {
-			outChunk.Column(r.outOrdinal).AppendManyNil(r.nilCount)
-			r.nilCount = 0
-		}
-	}
-}
-
-func (r *IntegerColFloatIntegralIterator) appendPrevItem(
-	inChunk, outChunk Chunk, tagIdx int,
-) {
-	outColumn := outChunk.Column(r.outOrdinal)
-	if r.isSingleCall {
-		for j := range r.buf.time {
-			outChunk.AppendTime(r.buf.time[j])
-			outColumn.AppendFloatValue(r.buf.value[j])
-			outColumn.AppendNotNil()
-		}
-		outChunk.AppendIntervalIndex(outChunk.Len() - r.buf.Len())
-		if !(outChunk.TagLen() > 0 && bytes.Equal(inChunk.Tags()[tagIdx].Subset(r.buf.opt.Dimensions),
-			outChunk.Tags()[outChunk.TagLen()-1].Subset(r.buf.opt.Dimensions))) {
-			outChunk.AppendTagsAndIndex(inChunk.Tags()[tagIdx], outChunk.IntervalIndex()[outChunk.IntervalLen()-1])
-		}
-		return
-	}
-
-	for j := range r.buf.time {
-		outColumn.AppendFloatValue(r.buf.value[j])
-		outColumn.AppendNotNil()
-	}
-}
-
-func (r *IntegerColFloatIntegralIterator) appendLastItem(
-	inChunk, outChunk Chunk, tagIdx int,
-) {
-	outColumn := outChunk.Column(r.outOrdinal)
-	if r.isSingleCall {
-		if r.buf.opt.Interval.IsZero() {
-			outChunk.AppendTime(0)
-		} else {
-			outChunk.AppendTime(r.buf.window.start)
-		}
-		outColumn.AppendFloatValue(r.buf.sum)
-		outColumn.AppendNotNil()
-		outChunk.AppendIntervalIndex(outChunk.Len() - 1)
-		if !(outChunk.TagLen() > 0 && bytes.Equal(inChunk.Tags()[tagIdx].Subset(r.buf.opt.Dimensions),
-			outChunk.Tags()[outChunk.TagLen()-1].Subset(r.buf.opt.Dimensions))) {
-			outChunk.AppendTagsAndIndex(inChunk.Tags()[tagIdx], outChunk.IntervalIndex()[outChunk.IntervalLen()-1])
-		}
-		return
-	}
-
-	outColumn.AppendFloatValue(r.buf.sum)
-	outColumn.AppendNotNil()
-}
-
-func (r *IntegerColFloatIntegralIterator) Next(ie *IteratorEndpoint, p *IteratorParams) {
-	inChunk, outChunk := ie.InputPoint.Chunk, ie.OutputPoint.Chunk
-	inColumn := inChunk.Column(r.inOrdinal)
-	if inColumn.IsEmpty() && r.buf.Len() > 0 {
-		return
-	}
-
-	var start, end int
-	firstIndex, lastIndex := 0, len(inChunk.IntervalIndex())-1
-	lastTagIndex := len(inChunk.TagIndex()) - 1
-	intervali := 0
-	var subIntervalIndexEnd int
-	outColumn := outChunk.Column(r.outOrdinal)
-	for i, _ := range inChunk.TagIndex() {
-		if i == lastTagIndex {
-			subIntervalIndexEnd = inChunk.NumberOfRows()
-		} else {
-			subIntervalIndexEnd = inChunk.TagIndex()[i+1]
-		}
-		for inChunk.IntervalIndex()[intervali] < subIntervalIndexEnd {
-			start = inChunk.IntervalIndex()[intervali]
-			if intervali < lastIndex {
-				end = inChunk.IntervalIndex()[intervali+1]
-			} else {
-				end = subIntervalIndexEnd
-			}
-			var sametag bool
-			if end == inChunk.NumberOfRows() {
-				sametag = !(end == subIntervalIndexEnd) || p.sameTag
-			} else {
-				sametag = !(end == subIntervalIndexEnd)
-			}
-			if !r.isSingleCall {
-				start, end = inColumn.GetRangeValueIndexV2(start, end)
-				if start == end {
-					if r.buf.Nil() {
-						outColumn.AppendNil()
-						intervali++
-						if intervali >= len(inChunk.IntervalIndex()) {
-							break
-						}
-						continue
-					} else {
-						r.nilCount++
-					}
-				}
-			}
-			if intervali == firstIndex && r.buf.sameInterval {
-				r.processFirstWindow(inChunk, firstIndex != lastIndex, p.sameInterval, sametag, start, end)
-			} else if intervali == lastIndex && p.sameInterval {
-				r.processLastWindow(inChunk, outChunk, start, end, i)
-			} else {
-				r.processMiddleWindow(inChunk, outChunk, start, end, i, sametag)
-			}
-			intervali++
-			if intervali >= len(inChunk.IntervalIndex()) {
-				break
-			}
-		}
-	}
-	if p.lastChunk {
-		if !(r.buf.pointNum == 1 && r.buf.prev.time == r.buf.window.start) && !(r.buf.pointNum == 0) {
-			r.appendLastItem(inChunk, outChunk, inChunk.TagLen()-1)
-		}
-		if !r.isSingleCall {
-			outColumn.AppendManyNil(r.nilCount)
-			r.nilCount = 0
-		}
-	}
-}
-
 type ElapsedItem struct {
 	prev     int64
 	time     []int64
@@ -5562,865 +4798,6 @@ func (r *BooleanColBooleanSampleIterator) Next(ie *IteratorEndpoint, p *Iterator
 			r.processLastWindow(inChunk, start, end)
 		} else {
 			r.processMiddleWindow(inChunk, outChunk, start, end)
-		}
-	}
-}
-
-type FloatSlidingWindowFloatIterator struct {
-	slidingNum int
-	inOrdinal  int
-	outOrdinal int
-	prevWindow *SlidingWindow[float64]
-	currWindow *SlidingWindow[float64]
-	fwr        ColReduceFunc[float64]
-	fpm        PointMerge[float64]
-	fwm        WindowMerge[float64]
-}
-
-func NewFloatSlidingWindowFloatIterator(
-	fwr ColReduceFunc[float64],
-	fpm PointMerge[float64],
-	fwm WindowMerge[float64],
-	inOrdinal, outOrdinal int, slidingNum int,
-) *FloatSlidingWindowFloatIterator {
-	r := &FloatSlidingWindowFloatIterator{
-		fwr:        fwr,
-		fpm:        fpm,
-		fwm:        fwm,
-		slidingNum: slidingNum,
-		inOrdinal:  inOrdinal,
-		outOrdinal: outOrdinal,
-		prevWindow: NewSlidingWindow[float64](slidingNum),
-		currWindow: NewSlidingWindow[float64](slidingNum),
-	}
-	return r
-}
-
-func (r *FloatSlidingWindowFloatIterator) mergePrevWindow(
-	outChunk Chunk,
-) {
-	outColumn := outChunk.Column(r.outOrdinal)
-	for i := 0; i < r.slidingNum; i++ {
-		if r.prevWindow.points[i].isNil {
-			outColumn.AppendNil()
-		} else {
-			outColumn.AppendNotNil()
-			outColumn.AppendFloatValue(r.prevWindow.points[i].value)
-		}
-	}
-}
-
-func (r *FloatSlidingWindowFloatIterator) processFirstWindow(
-	outChunk Chunk, value float64, isNil, sameTag, onlyOneWindow bool, n int,
-) {
-	r.currWindow.SetPoint(value, isNil, n)
-	if n < r.slidingNum-1 {
-		return
-	}
-	r.fwm(r.prevWindow, r.currWindow, r.fpm)
-	if !onlyOneWindow || !sameTag {
-		r.mergePrevWindow(outChunk)
-		r.prevWindow.Reset()
-	}
-	r.currWindow.Reset()
-}
-
-func (r *FloatSlidingWindowFloatIterator) processLastWindow(
-	value float64, isNil bool, index int,
-) {
-	r.prevWindow.SetPoint(value, isNil, index)
-}
-
-func (r *FloatSlidingWindowFloatIterator) processMiddleWindow(
-	outChunk Chunk, value float64, isNil bool,
-) {
-	outColumn := outChunk.Column(r.outOrdinal)
-	if isNil {
-		outColumn.AppendNil()
-	} else {
-		outColumn.AppendNotNil()
-		outColumn.AppendFloatValue(value)
-	}
-}
-
-func (r *FloatSlidingWindowFloatIterator) Next(ie *IteratorEndpoint, p *IteratorParams) {
-	var (
-		start, end int
-		value      float64
-		isNil      bool
-	)
-	firstIndex, lastIndex := 0, len(p.winIdx)/r.slidingNum-1
-	inChunk, outChunk := ie.InputPoint.Chunk, ie.OutputPoint.Chunk
-	values := inChunk.Column(r.inOrdinal).FloatValues()
-	for i := range p.winIdx {
-		m, n := i/r.slidingNum, i%r.slidingNum
-		start, end = p.winIdx[i][0], p.winIdx[i][1]
-		if start == -1 || end == -1 || start >= end {
-			value, isNil = 0, true
-		} else {
-			_, value, isNil = r.fwr(inChunk, values, r.inOrdinal, start, end)
-		}
-		if m == firstIndex && !r.prevWindow.IsNil() {
-			r.processFirstWindow(outChunk, value, isNil,
-				p.sameTag, firstIndex == lastIndex, n)
-		} else if m == lastIndex && p.sameTag {
-			r.processLastWindow(value, isNil, n)
-		} else {
-			r.processMiddleWindow(outChunk, value, isNil)
-		}
-	}
-}
-
-type BooleanColBooleanWindowReduce func(c Chunk, values []bool, ordinal, start, end int) (index int, value bool, isNil bool)
-
-type BooleanSlidingWindowBooleanIterator struct {
-	slidingNum int
-	inOrdinal  int
-	outOrdinal int
-	prevWindow *SlidingWindow[bool]
-	currWindow *SlidingWindow[bool]
-	fwr        BooleanColBooleanWindowReduce
-	fpm        PointMerge[bool]
-	fwm        WindowMerge[bool]
-}
-
-func NewBooleanSlidingWindowBooleanIterator(
-	fwr BooleanColBooleanWindowReduce,
-	fpm PointMerge[bool],
-	fwm WindowMerge[bool],
-	inOrdinal, outOrdinal int, slidingNum int,
-) *BooleanSlidingWindowBooleanIterator {
-	r := &BooleanSlidingWindowBooleanIterator{
-		fwr:        fwr,
-		fpm:        fpm,
-		fwm:        fwm,
-		slidingNum: slidingNum,
-		inOrdinal:  inOrdinal,
-		outOrdinal: outOrdinal,
-		prevWindow: NewSlidingWindow[bool](slidingNum),
-		currWindow: NewSlidingWindow[bool](slidingNum),
-	}
-	return r
-}
-
-func (r *BooleanSlidingWindowBooleanIterator) mergePrevWindow(
-	outChunk Chunk,
-) {
-	outColumn := outChunk.Column(r.outOrdinal)
-	for i := 0; i < r.slidingNum; i++ {
-		if r.prevWindow.points[i].isNil {
-			outColumn.AppendNil()
-		} else {
-			outColumn.AppendNotNil()
-			outColumn.AppendBooleanValue(r.prevWindow.points[i].value)
-		}
-	}
-}
-
-func (r *BooleanSlidingWindowBooleanIterator) processFirstWindow(
-	outChunk Chunk, value bool, isNil, sameTag, onlyOneWindow bool, n int,
-) {
-	r.currWindow.SetPoint(value, isNil, n)
-	if n < r.slidingNum-1 {
-		return
-	}
-	r.fwm(r.prevWindow, r.currWindow, r.fpm)
-	if !onlyOneWindow || !sameTag {
-		r.mergePrevWindow(outChunk)
-		r.prevWindow.Reset()
-	}
-	r.currWindow.Reset()
-}
-
-func (r *BooleanSlidingWindowBooleanIterator) processLastWindow(
-	value bool, isNil bool, index int,
-) {
-	r.prevWindow.SetPoint(value, isNil, index)
-}
-
-func (r *BooleanSlidingWindowBooleanIterator) processMiddleWindow(
-	outChunk Chunk, value bool, isNil bool,
-) {
-	outColumn := outChunk.Column(r.outOrdinal)
-	if isNil {
-		outColumn.AppendNil()
-	} else {
-		outColumn.AppendNotNil()
-		outColumn.AppendBooleanValue(value)
-	}
-}
-
-func (r *BooleanSlidingWindowBooleanIterator) Next(ie *IteratorEndpoint, p *IteratorParams) {
-	var (
-		start, end int
-		value      bool
-		isNil      bool
-	)
-	firstIndex, lastIndex := 0, len(p.winIdx)/r.slidingNum-1
-	inChunk, outChunk := ie.InputPoint.Chunk, ie.OutputPoint.Chunk
-	values := inChunk.Column(r.inOrdinal).BooleanValues()
-	for i := range p.winIdx {
-		m, n := i/r.slidingNum, i%r.slidingNum
-		start, end = p.winIdx[i][0], p.winIdx[i][1]
-		if start == -1 || end == -1 || start >= end {
-			value, isNil = false, true
-		} else {
-			_, value, isNil = r.fwr(inChunk, values, r.inOrdinal, start, end)
-		}
-		if m == firstIndex && !r.prevWindow.IsNil() {
-			r.processFirstWindow(outChunk, value, isNil,
-				p.sameTag, firstIndex == lastIndex, n)
-		} else if m == lastIndex && p.sameTag {
-			r.processLastWindow(value, isNil, n)
-		} else {
-			r.processMiddleWindow(outChunk, value, isNil)
-		}
-	}
-}
-
-type IntegerColIntegerWindowReduce func(c Chunk, ordinal, start, end int) (index int, value int64, isNil bool)
-
-type IntegerSlidingWindowIntegerIterator struct {
-	slidingNum int
-	inOrdinal  int
-	outOrdinal int
-	prevWindow *SlidingWindow[int64]
-	currWindow *SlidingWindow[int64]
-	fwr        ColReduceFunc[int64]
-	fpm        PointMerge[int64]
-	fwm        WindowMerge[int64]
-}
-
-func NewIntegerSlidingWindowIntegerIterator(
-	fwr ColReduceFunc[int64],
-	fpm PointMerge[int64],
-	fwm WindowMerge[int64],
-	inOrdinal, outOrdinal int, slidingNum int,
-) *IntegerSlidingWindowIntegerIterator {
-	r := &IntegerSlidingWindowIntegerIterator{
-		fwr:        fwr,
-		fpm:        fpm,
-		fwm:        fwm,
-		slidingNum: slidingNum,
-		inOrdinal:  inOrdinal,
-		outOrdinal: outOrdinal,
-		prevWindow: NewSlidingWindow[int64](slidingNum),
-		currWindow: NewSlidingWindow[int64](slidingNum),
-	}
-	return r
-}
-
-func (r *IntegerSlidingWindowIntegerIterator) mergePrevWindow(
-	outChunk Chunk,
-) {
-	outColumn := outChunk.Column(r.outOrdinal)
-	for i := 0; i < r.slidingNum; i++ {
-		if r.prevWindow.points[i].isNil {
-			outColumn.AppendNil()
-		} else {
-			outColumn.AppendNotNil()
-			outColumn.AppendIntegerValue(r.prevWindow.points[i].value)
-		}
-	}
-}
-
-func (r *IntegerSlidingWindowIntegerIterator) processFirstWindow(
-	outChunk Chunk, value int64, isNil, sameTag, onlyOneWindow bool, n int,
-) {
-	r.currWindow.SetPoint(value, isNil, n)
-	if n < r.slidingNum-1 {
-		return
-	}
-	r.fwm(r.prevWindow, r.currWindow, r.fpm)
-	if !onlyOneWindow || !sameTag {
-		r.mergePrevWindow(outChunk)
-		r.prevWindow.Reset()
-	}
-	r.currWindow.Reset()
-}
-
-func (r *IntegerSlidingWindowIntegerIterator) processLastWindow(
-	value int64, isNil bool, index int,
-) {
-	r.prevWindow.SetPoint(value, isNil, index)
-}
-
-func (r *IntegerSlidingWindowIntegerIterator) processMiddleWindow(
-	outChunk Chunk, value int64, isNil bool,
-) {
-	outColumn := outChunk.Column(r.outOrdinal)
-	if isNil {
-		outColumn.AppendNil()
-	} else {
-		outColumn.AppendNotNil()
-		outColumn.AppendIntegerValue(value)
-	}
-}
-
-func (r *IntegerSlidingWindowIntegerIterator) Next(ie *IteratorEndpoint, p *IteratorParams) {
-	var (
-		start, end int
-		value      int64
-		isNil      bool
-	)
-	firstIndex, lastIndex := 0, len(p.winIdx)/r.slidingNum-1
-	inChunk, outChunk := ie.InputPoint.Chunk, ie.OutputPoint.Chunk
-	values := inChunk.Column(r.inOrdinal).IntegerValues()
-	for i := range p.winIdx {
-		m, n := i/r.slidingNum, i%r.slidingNum
-		start, end = p.winIdx[i][0], p.winIdx[i][1]
-		if start == -1 || end == -1 || start >= end {
-			value, isNil = 0, true
-		} else {
-			_, value, isNil = r.fwr(inChunk, values, r.inOrdinal, start, end)
-		}
-		if m == firstIndex && !r.prevWindow.IsNil() {
-			r.processFirstWindow(outChunk, value, isNil,
-				p.sameTag, firstIndex == lastIndex, n)
-		} else if m == lastIndex && p.sameTag {
-			r.processLastWindow(value, isNil, n)
-		} else {
-			r.processMiddleWindow(outChunk, value, isNil)
-		}
-	}
-}
-
-type BooleanColIntegerWindowReduce func(c Chunk, ordinal, start, end int) (index int, value int64, isNil bool)
-
-type BooleanSlidingWindowIntegerIterator struct {
-	slidingNum int
-	inOrdinal  int
-	outOrdinal int
-	prevWindow *SlidingWindow[int64]
-	currWindow *SlidingWindow[int64]
-	fwr        ColReduceFunc[int64]
-	fpm        PointMerge[int64]
-	fwm        WindowMerge[int64]
-}
-
-func NewBooleanSlidingWindowIntegerIterator(
-	fwr ColReduceFunc[int64],
-	fpm PointMerge[int64],
-	fwm WindowMerge[int64],
-	inOrdinal, outOrdinal int, slidingNum int,
-) *BooleanSlidingWindowIntegerIterator {
-
-	r := &BooleanSlidingWindowIntegerIterator{
-		fwr:        fwr,
-		fpm:        fpm,
-		fwm:        fwm,
-		slidingNum: slidingNum,
-		inOrdinal:  inOrdinal,
-		outOrdinal: outOrdinal,
-		prevWindow: NewSlidingWindow[int64](slidingNum),
-		currWindow: NewSlidingWindow[int64](slidingNum),
-	}
-	return r
-}
-
-func (r *BooleanSlidingWindowIntegerIterator) mergePrevWindow(
-	outChunk Chunk,
-) {
-	outColumn := outChunk.Column(r.outOrdinal)
-	for i := 0; i < r.slidingNum; i++ {
-		if r.prevWindow.points[i].isNil {
-			outColumn.AppendNil()
-		} else {
-			outColumn.AppendNotNil()
-			outColumn.AppendIntegerValue(r.prevWindow.points[i].value)
-		}
-	}
-}
-
-func (r *BooleanSlidingWindowIntegerIterator) processFirstWindow(
-	outChunk Chunk, value int64, isNil, sameTag, onlyOneWindow bool, n int,
-) {
-	r.currWindow.SetPoint(value, isNil, n)
-	if n < r.slidingNum-1 {
-		return
-	}
-	r.fwm(r.prevWindow, r.currWindow, r.fpm)
-	if !onlyOneWindow || !sameTag {
-		r.mergePrevWindow(outChunk)
-		r.prevWindow.Reset()
-	}
-	r.currWindow.Reset()
-}
-
-func (r *BooleanSlidingWindowIntegerIterator) processLastWindow(
-	value int64, isNil bool, index int,
-) {
-	r.prevWindow.SetPoint(value, isNil, index)
-}
-
-func (r *BooleanSlidingWindowIntegerIterator) processMiddleWindow(
-	outChunk Chunk, value int64, isNil bool,
-) {
-	outColumn := outChunk.Column(r.outOrdinal)
-	if isNil {
-		outColumn.AppendNil()
-	} else {
-		outColumn.AppendNotNil()
-		outColumn.AppendIntegerValue(value)
-	}
-}
-
-func (r *BooleanSlidingWindowIntegerIterator) Next(ie *IteratorEndpoint, p *IteratorParams) {
-	var (
-		start, end int
-		value      int64
-		isNil      bool
-	)
-	firstIndex, lastIndex := 0, len(p.winIdx)/r.slidingNum-1
-	inChunk, outChunk := ie.InputPoint.Chunk, ie.OutputPoint.Chunk
-	values := inChunk.Column(r.inOrdinal).IntegerValues()
-	for i := range p.winIdx {
-		m, n := i/r.slidingNum, i%r.slidingNum
-		start, end = p.winIdx[i][0], p.winIdx[i][1]
-		if start == -1 || end == -1 || start >= end {
-			value, isNil = 0, true
-		} else {
-			_, value, isNil = r.fwr(inChunk, values, r.inOrdinal, start, end)
-		}
-		if m == firstIndex && !r.prevWindow.IsNil() {
-			r.processFirstWindow(outChunk, value, isNil,
-				p.sameTag, firstIndex == lastIndex, n)
-		} else if m == lastIndex && p.sameTag {
-			r.processLastWindow(value, isNil, n)
-		} else {
-			r.processMiddleWindow(outChunk, value, isNil)
-		}
-	}
-}
-
-type OGSketchItem interface {
-	UpdateCluster(inChunk Chunk, start, end int)
-	WriteResult(outChunk Chunk, time int64)
-	IsNil() bool
-	Reset()
-}
-
-type FloatOGSketchInsertItem struct {
-	isSingleCall bool
-	inOrdinal    int
-	outOrdinal   int
-	clusterNum   int
-	percentile   float64
-	sketch       OGSketch
-	clusters     ClusterSet
-}
-
-func NewFloatOGSketchInsertIem(isSingleCall bool, inOrdinal, outOrdinal, clusterNum int, percentile float64) *FloatOGSketchInsertItem {
-	return &FloatOGSketchInsertItem{
-		isSingleCall: isSingleCall,
-		inOrdinal:    inOrdinal,
-		outOrdinal:   outOrdinal,
-		clusterNum:   clusterNum,
-		percentile:   percentile,
-		sketch:       NewOGSketchImpl(float64(clusterNum)),
-	}
-}
-
-func (o *FloatOGSketchInsertItem) UpdateCluster(inChunk Chunk, start, end int) {
-	o.sketch.InsertPoints(inChunk.Column(o.inOrdinal).FloatValues()[start:end]...)
-}
-
-func (o *FloatOGSketchInsertItem) WriteResult(outChunk Chunk, time int64) {
-	o.clusters = o.sketch.Clusters()
-	clusterNum := len(o.clusters)
-	if o.isSingleCall {
-		for i := 0; i < clusterNum; i++ {
-			outChunk.AppendTime(time)
-		}
-		outChunk.AppendIntervalIndex(outChunk.Len() - clusterNum)
-	}
-	outColumn := outChunk.Column(o.outOrdinal)
-	for i := 0; i < clusterNum; i++ {
-		outColumn.AppendFloatTuple(floatTuple{values: []float64{o.clusters[i].Mean, o.clusters[i].Weight}})
-	}
-	outColumn.AppendManyNotNil(clusterNum)
-
-	if !o.isSingleCall && o.clusterNum > clusterNum {
-		outColumn.AppendManyNil(o.clusterNum - clusterNum)
-	}
-}
-
-func (o *FloatOGSketchInsertItem) IsNil() bool {
-	return o.sketch.Len() == 0
-}
-
-func (o *FloatOGSketchInsertItem) Reset() {
-	o.sketch.Reset()
-}
-
-type FloatOGSketchPercentileItem struct {
-	isSingleCall bool
-	inOrdinal    int
-	outOrdinal   int
-	clusterNum   int
-	percentile   float64
-	sketch       OGSketch
-}
-
-func NewFloatOGSketchPercentileItem(isSingleCall bool, inOrdinal, outOrdinal, clusterNum int, percentile float64) *FloatOGSketchPercentileItem {
-	return &FloatOGSketchPercentileItem{
-		isSingleCall: isSingleCall,
-		inOrdinal:    inOrdinal,
-		outOrdinal:   outOrdinal,
-		clusterNum:   clusterNum,
-		percentile:   percentile,
-		sketch:       NewOGSketchImpl(float64(clusterNum)),
-	}
-}
-
-func (o *FloatOGSketchPercentileItem) UpdateCluster(inChunk Chunk, start, end int) {
-	tuples := inChunk.Column(o.inOrdinal).FloatTuples()[start:end]
-	o.sketch.InsertClusters(tuples...)
-
-}
-
-func (o *FloatOGSketchPercentileItem) WriteResult(outChunk Chunk, time int64) {
-	value := o.sketch.Percentile(o.percentile)
-	if o.isSingleCall {
-		outChunk.AppendTime(time)
-		outChunk.AppendIntervalIndex(outChunk.Len() - 1)
-	}
-	outColumn := outChunk.Column(o.outOrdinal)
-	outColumn.AppendFloatValue(value)
-	outColumn.AppendNotNil()
-}
-
-func (o *FloatOGSketchPercentileItem) IsNil() bool {
-	return o.sketch.Len() == 0
-}
-
-func (o *FloatOGSketchPercentileItem) Reset() {
-	o.sketch.Reset()
-}
-
-type FloatPercentileApproxItem struct {
-	isSingleCall bool
-	inOrdinal    int
-	outOrdinal   int
-	clusterNum   int
-	percentile   float64
-	sketch       OGSketch
-}
-
-func NewFloatPercentileApproxItem(isSingleCall bool, inOrdinal, outOrdinal, clusterNum int, percentile float64) *FloatPercentileApproxItem {
-	return &FloatPercentileApproxItem{
-		isSingleCall: isSingleCall,
-		inOrdinal:    inOrdinal,
-		outOrdinal:   outOrdinal,
-		clusterNum:   clusterNum,
-		percentile:   percentile,
-		sketch:       NewOGSketchImpl(float64(clusterNum)),
-	}
-}
-
-func (o *FloatPercentileApproxItem) UpdateCluster(inChunk Chunk, start, end int) {
-	o.sketch.InsertPoints(inChunk.Column(o.inOrdinal).FloatValues()[start:end]...)
-}
-
-func (o *FloatPercentileApproxItem) WriteResult(outChunk Chunk, time int64) {
-	value := o.sketch.Percentile(o.percentile)
-	if o.isSingleCall {
-		outChunk.AppendTime(time)
-		outChunk.AppendIntervalIndex(outChunk.Len() - 1)
-	}
-	outColumn := outChunk.Column(o.outOrdinal)
-	outColumn.AppendFloatValue(value)
-	outColumn.AppendNotNil()
-}
-
-func (o *FloatPercentileApproxItem) IsNil() bool {
-	return o.sketch.Len() == 0
-}
-
-func (o *FloatPercentileApproxItem) Reset() {
-	o.sketch.Reset()
-}
-
-type IntegerOGSketchInsertItem struct {
-	isSingleCall bool
-	inOrdinal    int
-	outOrdinal   int
-	clusterNum   int
-	percentile   float64
-	sketch       OGSketch
-	clusters     ClusterSet
-}
-
-func NewIntegerOGSketchInsertIem(isSingleCall bool, inOrdinal, outOrdinal, clusterNum int, percentile float64) *IntegerOGSketchInsertItem {
-	return &IntegerOGSketchInsertItem{
-		isSingleCall: isSingleCall,
-		inOrdinal:    inOrdinal,
-		outOrdinal:   outOrdinal,
-		clusterNum:   clusterNum,
-		percentile:   percentile,
-		sketch:       NewOGSketchImpl(float64(clusterNum)),
-	}
-}
-
-func (o *IntegerOGSketchInsertItem) UpdateCluster(inChunk Chunk, start, end int) {
-	values := inChunk.Column(o.inOrdinal).IntegerValues()[start:end]
-	for i := range values {
-		o.sketch.InsertPoints(float64(values[i]))
-	}
-}
-
-func (o *IntegerOGSketchInsertItem) WriteResult(outChunk Chunk, time int64) {
-	o.clusters = o.sketch.Clusters()
-	clusterNum := len(o.clusters)
-	if o.isSingleCall {
-		for i := 0; i < clusterNum; i++ {
-			outChunk.AppendTime(time)
-		}
-		outChunk.AppendIntervalIndex(outChunk.Len() - clusterNum)
-	}
-
-	outColumn := outChunk.Column(o.outOrdinal)
-	for i := 0; i < clusterNum; i++ {
-		outColumn.AppendFloatTuple(floatTuple{values: []float64{o.clusters[i].Mean, o.clusters[i].Weight}})
-	}
-	outColumn.AppendManyNotNil(clusterNum)
-
-	if !o.isSingleCall && o.clusterNum > clusterNum {
-		outColumn.AppendManyNil(o.clusterNum - clusterNum)
-	}
-}
-
-func (o *IntegerOGSketchInsertItem) IsNil() bool {
-	return o.sketch.Len() == 0
-}
-
-func (o *IntegerOGSketchInsertItem) Reset() {
-	o.sketch.Reset()
-}
-
-type IntegerOGSketchPercentileItem struct {
-	isSingleCall bool
-	inOrdinal    int
-	outOrdinal   int
-	clusterNum   int
-	percentile   float64
-	sketch       OGSketch
-}
-
-func NewIntegerOGSketchPercentileItem(isSingleCall bool, inOrdinal, outOrdinal, clusterNum int, percentile float64) *IntegerOGSketchPercentileItem {
-	return &IntegerOGSketchPercentileItem{
-		isSingleCall: isSingleCall,
-		inOrdinal:    inOrdinal,
-		outOrdinal:   outOrdinal,
-		clusterNum:   clusterNum,
-		percentile:   percentile,
-		sketch:       NewOGSketchImpl(float64(clusterNum)),
-	}
-}
-
-func (o *IntegerOGSketchPercentileItem) UpdateCluster(inChunk Chunk, start, end int) {
-	tuples := inChunk.Column(o.inOrdinal).FloatTuples()[start:end]
-	o.sketch.InsertClusters(tuples...)
-
-}
-
-func (o *IntegerOGSketchPercentileItem) WriteResult(outChunk Chunk, time int64) {
-	value := o.sketch.Percentile(o.percentile)
-	if o.isSingleCall {
-		outChunk.AppendTime(time)
-		outChunk.AppendIntervalIndex(outChunk.Len() - 1)
-	}
-	outColumn := outChunk.Column(o.outOrdinal)
-	outColumn.AppendIntegerValue(int64(value))
-	outColumn.AppendNotNil()
-}
-
-func (o *IntegerOGSketchPercentileItem) IsNil() bool {
-	return o.sketch.Len() == 0
-}
-
-func (o *IntegerOGSketchPercentileItem) Reset() {
-	o.sketch.Reset()
-}
-
-type IntegerPercentileApproxItem struct {
-	isSingleCall bool
-	inOrdinal    int
-	outOrdinal   int
-	clusterNum   int
-	percentile   float64
-	sketch       OGSketch
-}
-
-func NewIntegerPercentileApproxItem(isSingleCall bool, inOrdinal, outOrdinal, clusterNum int, percentile float64) *IntegerPercentileApproxItem {
-	return &IntegerPercentileApproxItem{
-		isSingleCall: isSingleCall,
-		inOrdinal:    inOrdinal,
-		outOrdinal:   outOrdinal,
-		clusterNum:   clusterNum,
-		percentile:   percentile,
-		sketch:       NewOGSketchImpl(float64(clusterNum)),
-	}
-}
-
-func (o *IntegerPercentileApproxItem) UpdateCluster(inChunk Chunk, start, end int) {
-	values := inChunk.Column(o.inOrdinal).IntegerValues()[start:end]
-	for i := range values {
-		o.sketch.InsertPoints(float64(values[i]))
-	}
-}
-
-func (o *IntegerPercentileApproxItem) WriteResult(outChunk Chunk, time int64) {
-	value := o.sketch.Percentile(o.percentile)
-	if o.isSingleCall {
-		outChunk.AppendTime(time)
-		outChunk.AppendIntervalIndex(outChunk.Len() - 1)
-	}
-	outColumn := outChunk.Column(o.outOrdinal)
-	outColumn.AppendIntegerValue(int64(value))
-	outColumn.AppendNotNil()
-}
-
-func (o *IntegerPercentileApproxItem) IsNil() bool {
-	return o.sketch.Len() == 0
-}
-
-func (o *IntegerPercentileApproxItem) Reset() {
-	o.sketch.Reset()
-}
-
-type OGSketchMergeItem struct {
-	isSingleCall bool
-	inOrdinal    int
-	outOrdinal   int
-	clusterNum   int
-	percentile   float64
-	sketch       OGSketch
-	clusters     ClusterSet
-}
-
-func NewOGSketchMergeItem(isSingleCall bool, inOrdinal, outOrdinal, clusterNum int, percentile float64) *OGSketchMergeItem {
-	return &OGSketchMergeItem{
-		isSingleCall: isSingleCall,
-		inOrdinal:    inOrdinal,
-		outOrdinal:   outOrdinal,
-		clusterNum:   clusterNum,
-		percentile:   percentile,
-		sketch:       NewOGSketchImpl(float64(clusterNum)),
-	}
-}
-
-func (o *OGSketchMergeItem) UpdateCluster(inChunk Chunk, start, end int) {
-	tuples := inChunk.Column(o.inOrdinal).FloatTuples()[start:end]
-	o.sketch.InsertClusters(tuples...)
-}
-
-func (o *OGSketchMergeItem) WriteResult(outChunk Chunk, time int64) {
-	o.clusters = o.sketch.Clusters()
-	clusterNum := len(o.clusters)
-	if o.isSingleCall {
-		for i := 0; i < clusterNum; i++ {
-			outChunk.AppendTime(time)
-		}
-		outChunk.AppendIntervalIndex(outChunk.Len() - clusterNum)
-	}
-	outColumn := outChunk.Column(o.outOrdinal)
-	for i := 0; i < clusterNum; i++ {
-		outColumn.AppendFloatTuple(floatTuple{values: []float64{o.clusters[i].Mean, o.clusters[i].Weight}})
-	}
-	outColumn.AppendManyNotNil(clusterNum)
-
-	if !o.isSingleCall && o.clusterNum > clusterNum {
-		outColumn.AppendManyNil(o.clusterNum - clusterNum)
-	}
-}
-
-func (o *OGSketchMergeItem) IsNil() bool {
-	return o.sketch.Len() == 0
-}
-
-func (o *OGSketchMergeItem) Reset() {
-	o.sketch.Reset()
-}
-
-type OGSketchIterator struct {
-	isSingleCall bool
-	inOrdinal    int
-	outOrdinal   int
-	clusterNum   int
-	opt          *query.ProcessorOptions
-	sketch       OGSketchItem
-}
-
-func NewOGSketchIterator(
-	isSingleCall bool, inOrdinal, outOrdinal int, clusterNum int, opt *query.ProcessorOptions, sketch OGSketchItem,
-) *OGSketchIterator {
-	r := &OGSketchIterator{
-		isSingleCall: isSingleCall,
-		inOrdinal:    inOrdinal,
-		outOrdinal:   outOrdinal,
-		clusterNum:   clusterNum,
-		opt:          opt,
-		sketch:       sketch,
-	}
-	return r
-}
-
-func (r *OGSketchIterator) processFirstWindow(
-	inChunk, outChunk Chunk, sameInterval, haveMultiInterval bool, start, end int, time int64,
-) {
-	r.sketch.UpdateCluster(inChunk, start, end)
-	if haveMultiInterval || !sameInterval {
-		r.sketch.WriteResult(outChunk, time)
-		r.sketch.Reset()
-	}
-}
-
-func (r *OGSketchIterator) processLastWindow(
-	inChunk Chunk, start, end int,
-) {
-	r.sketch.UpdateCluster(inChunk, start, end)
-}
-
-func (r *OGSketchIterator) processMiddleWindow(
-	inChunk, outChunk Chunk, start, end int, time int64,
-) {
-	r.sketch.UpdateCluster(inChunk, start, end)
-	r.sketch.WriteResult(outChunk, time)
-	r.sketch.Reset()
-}
-
-func (r *OGSketchIterator) Next(ie *IteratorEndpoint, p *IteratorParams) {
-	var end int
-	inChunk, outChunk := ie.InputPoint.Chunk, ie.OutputPoint.Chunk
-	firstIndex, lastIndex := 0, len(inChunk.IntervalIndex())-1
-	inColumn := inChunk.Column(r.inOrdinal)
-	outColumn := outChunk.Column(r.outOrdinal)
-	for i, start := range inChunk.IntervalIndex() {
-		if i < lastIndex {
-			end = inChunk.IntervalIndex()[i+1]
-		} else {
-			end = inChunk.NumberOfRows()
-		}
-		var time int64
-		if r.opt.Interval.IsZero() {
-			time = hybridqp.MaxInt64(r.opt.StartTime, 0)
-		} else {
-			time, _ = r.opt.Window(inChunk.TimeByIndex(start))
-		}
-		start, end = inColumn.GetRangeValueIndexV2(start, end)
-		if !r.isSingleCall {
-			if start == end && r.sketch.IsNil() && (i < lastIndex || (i == lastIndex && !p.sameInterval)) {
-				outColumn.AppendManyNil(r.clusterNum)
-				continue
-			}
-		}
-		if i == firstIndex && !r.sketch.IsNil() {
-			r.processFirstWindow(inChunk, outChunk, p.sameInterval, firstIndex != lastIndex, start, end, time)
-		} else if i == lastIndex && p.sameInterval {
-			r.processLastWindow(inChunk, start, end)
-		} else {
-			r.processMiddleWindow(inChunk, outChunk, start, end, time)
 		}
 	}
 }
